Move CommunicationUtilities_Tests.cs to Framework.UnitTests
